fix refs to core types
