+++ crates/network/src/types/message.rs
@@ -245,7 +245,7 @@ pub enum MessageBody {
     },
     AddPeer {
         // pub_key: PublicKey,
-        peer_id: primitives::types::PeerId,
+        peer_id: primitives::PeerId,
         socket_addr: SocketAddr,
         node_type: NodeType,
     },
