+++ crates/primitives/src/lib.rs
@@ -1,5 +1,13 @@
+pub mod address;
+pub mod base;
+pub mod crypto;
+pub mod digest;
 pub mod node;
-pub mod types;
+pub mod signal;
 
+pub use address::*;
+pub use base::*;
+pub use crypto::*;
+pub use digest::*;
 pub use node::*;
-pub use types::*;
+pub use signal::*;
