+++ crates/primitives/src/signal.rs
@@ -1,0 +1,2 @@
+#[derive(Clone, Debug, Default)]
+pub struct StopSignal;
