+++ crates/validator/src/txn_validator.rs
@@ -7,7 +7,6 @@ use std::{
 use left_right::ReadHandle;
 use lr_trie::LeftRightTrieError;
 use patriecia::{db::Database, error::TrieError, inner::InnerTrie};
-use primitives::types::AccountAddress;
 use vrrb_core::{
     account::Account,
     keypair::{KeyPair, MinerPk},
